\name{Basingstoke}
\alias{Basingstoke}
\alias{Basingstoke-package}
\docType{package}
\title{Food-web structural models}
\description{
Food-web structural models
}
\details{
\tabular{ll}{
    Package: \tab Basingstoke\cr
    Type: \tab Package\cr
    Version: \tab 0.1.0\cr
    Date: \tab 2015-04-15\cr
    License: \tab BSD 2 clause\cr
}
}

\author{Maintainer: Lawrence Hudson <l.hudson@nhm.ac.uk}

% \references{}

\keyword{ package }

% \examples{
% }
